backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
gunicorn>=21.0.0
supabase>=2.0.0
openai>=1.3.0
pydantic>=2.5.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10
newspaper3k>=0.2.8
apscheduler>=3.10.0
python-dateutil>=2.8.0
aiohttp>=3.9.0
httpx>=0.26.0
python-dotenv>=1.0.0

# Security middleware
slowapi>=0.1.9

# GPT Researcher for deep research functionality
gpt-researcher>=0.9.0
tavily-python>=0.3.0

# Export and reporting functionality
python-pptx>=0.6.21
reportlab>=4.0.0
pandas>=2.0.0
matplotlib>=3.8.0
trafilatura>=2.0.0

# SQLAlchemy + async PostgreSQL driver (Azure PG migration)
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# Azure Blob Storage for document attachments
azure-storage-blob>=12.19.0

# DOCX export
python-docx>=1.1.0
